fix: main.css, query, date format

+++ src/main/java/crafty/dto/ResponseGoodsManagement.java
@@ -14,5 +14,5 @@ public class ResponseGoodsManagement {
 	private String	nickname;
 	private int		amount;
 	private	Date	adjustmentCreatedAt;
-	private int		lastTotalAmount;
+	private Long	lastTotalAmount;
 }
